# api/openapi.yaml
openapi: 3.1.0
info:
  title: flexitype API
  description: >
    Runtime-defined soft types, attributes, dependencies, relationships and
    a query language (FQL), plus event delivery. All endpoints are tenant
    scoped via the authenticated service account. See docs/api-stability.md
    for compatibility guarantees.
  version: v1
  license:
    name: MIT
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /attributes:
    get:
      tags: [attributes]
      summary: List attribute definitions
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: type_definition_id
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/IncludeArchived"
      responses:
        "200": { $ref: "#/components/responses/AttributePage" }
    post:
      tags: [attributes]
      summary: Create an attribute definition
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAttribute" }
      responses:
        "201": { $ref: "#/components/responses/Attribute" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /schema/import:
    post:
      tags: [schema]
      summary: Import a schema bundle (idempotent)
      description: >
        Creates objects missing by internal name and skips those already
        present, in dependency order. Re-running is safe and completes a
        partial import. Not a single transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SchemaBundle" }
      responses:
        "200":
          description: Per-kind created/skipped counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  types: { $ref: "#/components/schemas/ImportKindCount" }
                  attributes: { $ref: "#/components/schemas/ImportKindCount" }
                  relationship_definitions: { $ref: "#/components/schemas/ImportKindCount" }
                  dependencies: { $ref: "#/components/schemas/ImportKindCount" }
        "422": { $ref: "#/components/responses/Error" }
...
  /query:
    get:
      tags: [query]
      summary: Run an FQL query
      parameters:
        - { name: type, in: query, required: true, schema: { type: string }, description: Root type internal name }
        - { name: q, in: query, required: true, schema: { type: string }, description: FQL expression }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Matching entities
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/EntitySummary" } }
                  page_info: { $ref: "#/components/schemas/PageInfo" }
        "422": { $ref: "#/components/responses/Error" }
        "501": { $ref: "#/components/responses/Error" }
...
  /entities/{typeDefinitionID}/import:
    parameters: [{ name: typeDefinitionID, in: path, required: true, schema: { type: string } }]
    post:
      tags: [entities]
      summary: Import a CSV of entity values
      description: >
        Multipart upload: the CSV as `file` and a JSON `mapping` describing the
        key column, column-to-attribute map, commit mode and dry-run flag.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, mapping]
              properties:
                file: { type: string, format: binary }
                mapping: { type: string, description: "JSON: key_column, mapping, mode, dry_run" }
      responses:
        "200": { description: Import report, content: { application/json: { schema: { $ref: "#/components/schemas/ImportReport" } } } }
        "422": { $ref: "#/components/responses/Error" }
  /entities/{typeDefinitionID}/export:
    parameters: [{ name: typeDefinitionID, in: path, required: true, schema: { type: string } }]
    get:
      tags: [entities]
      summary: Export a type's entities as CSV
      parameters:
        - { name: attributes, in: query, schema: { type: string }, description: Comma-separated attribute internal names (columns) }
        - { name: query, in: query, schema: { type: string }, description: Optional FQL filter }
        - { name: entity_ids, in: query, schema: { type: string }, description: Optional explicit comma-separated entity ids }
      responses:
        "200":
          description: CSV export
          content:
            text/csv:
              schema: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Service-account token, `ft_<account>_<secret>`. Omit auth in
        development mode (no service-account file configured).
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 500 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    IncludeArchived:
      name: include_archived
      in: query
      schema: { type: boolean }
  responses:
    Error:
      description: Error with a stable machine code
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Attribute:
      description: An attribute definition
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Attribute" }
    AttributePage:
      description: A page of attribute definitions
      content:
        application/json:
          schema:
            type: object
            properties:
              items: { type: array, items: { $ref: "#/components/schemas/Attribute" } }
              page_info: { $ref: "#/components/schemas/PageInfo" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION, NOT_FOUND, CONFLICT, ARCHIVED, DEPENDENCY_VIOLATION, FEATURE_DISABLED, CURSOR_CONFLICT, CURSOR_EXPIRED, UNAUTHENTICATED, FORBIDDEN, RATE_LIMITED, INTERNAL]
            message: { type: string }
            details: { type: object, additionalProperties: true }
    PageInfo:
      type: object
      properties:
        total_count: { type: integer }
        has_next_page: { type: boolean }
        has_previous_page: { type: boolean }
        next_cursor: { type: string }
    DataType:
      type: string
      enum: [bool, string, integer, float, decimal, date, time, datetime, enum, url, email, json]
...
    Attribute:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        type_definition_id: { type: string }
        internal_name: { type: string }
        display_name: { type: string }
        description: { type: string }
        data_type: { $ref: "#/components/schemas/DataType" }
        required: { type: boolean }
        multi_valued: { type: boolean }
        unique: { type: boolean }
        group: { type: string }
        sort_order: { type: integer }
        help_text: { type: string }
        constraints: { type: array, items: { type: object } }
        default_value: {}
        version: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        archived_at: { type: string, format: date-time }
    CreateAttribute:
      type: object
      required: [type_definition_id, internal_name, display_name, data_type]
      properties:
        type_definition_id: { type: string }
        internal_name: { type: string }
        display_name: { type: string }
        description: { type: string }
        data_type: { $ref: "#/components/schemas/DataType" }
        required: { type: boolean }
        multi_valued: { type: boolean }
        unique: { type: boolean }
        group: { type: string }
        sort_order: { type: integer }
        help_text: { type: string }
        constraints: { type: array, items: { type: object } }
        default_value: {}
...
    EntitySummary:
      type: object
      properties:
        entity_id: { type: string }
        type_definition_id: { type: string }
        value_count: { type: integer }
        last_updated_at: { type: string, format: date-time }
...
    ImportKindCount:
      type: object
      properties:
        created: { type: integer }
        skipped: { type: integer }
...
    SchemaBundle:
      type: object
      description: A portable, internal-name-keyed snapshot of a tenant's schema.
      required: [version]
      properties:
        version: { type: integer, description: Bundle format version (currently 1) }
        types:
          type: array
          items:
            type: object
            properties:
              internal_name: { type: string }
              display_name: { type: string }
              description: { type: string }
              extends: { type: string, description: Parent type internal name }
        attributes:
          type: array
          items:
            type: object
            properties:
              type: { type: string, description: Declaring type internal name }
              internal_name: { type: string }
              display_name: { type: string }
              description: { type: string }
              data_type: { $ref: "#/components/schemas/DataType" }
              required: { type: boolean }
              multi_valued: { type: boolean }
              unique: { type: boolean }
              constraints: { type: array, items: { type: object, additionalProperties: true } }
              default_value: { type: object, additionalProperties: true }
        relationship_definitions:
          type: array
          items:
            type: object
            properties:
              internal_name: { type: string }
              display_name: { type: string }
              description: { type: string }
              kind: { type: string, enum: [directed, symmetric] }
              parent_type: { type: string }
              child_type: { type: string }
              parent_label: { type: string }
              child_label: { type: string }
              parent_version_policy: { type: string }
              child_version_policy: { type: string }
        dependencies:
          type: array
          items:
            type: object
            properties:
              source_type: { type: string }
              source_attribute: { type: string }
              target_type: { type: string }
              target_attribute: { type: string }
              conditions: { type: array, items: { type: object, additionalProperties: true } }
              effect: { type: object, additionalProperties: true }
              description: { type: string }
...
    ImportReport:
      type: object
      properties:
        rows_total: { type: integer }
        rows_valid: { type: integer }
        rows_written: { type: integer }
        dry_run: { type: boolean }
        mode: { type: string }
        errors: { type: array, items: { $ref: "#/components/schemas/ImportError" } }
    ImportError:
      type: object
      properties:
        row: { type: integer }
        column: { type: string }
        attribute: { type: string }
        reason: { type: string }
...